+++ projects/project-001-vector-search/src/embeddings/embeddingService.js
@@ -1,0 +1,12 @@
+export function generateEmbedding(text){
+
+    // Simulação de um embedding.
+    // Na próxima aula este retorno será substituído
+    // pela chamada real da OpenAI.
+
+    return [
+        Math.random(),
+        Math.random(),
+        Math.random()
+    ];
+}